Moving extensions to their own class
