+++ include/simulator/base/component.h
@@ -18,12 +18,12 @@ class Component : public StatsOutputter {
     /* Update component with how much time has passed for operation. Used for updating static power. */
     virtual void update_static(double time) = 0;
     /* Returns the total amount of static energy use */
-    virtual cv::Mat& get_static_energy() = 0;
+    virtual cv::Mat get_static_energy() = 0;
     /* Returns the total amount of dynamic energy use */
-    virtual cv::Mat& get_dynamic_energy() = 0;
+    virtual cv::Mat get_dynamic_energy() = 0;
     /* Returns the transistor count across the array
      * This should be a constant after the component is configured. */
-    virtual cv::Mat& get_transistor_count() = 0;
+    virtual cv::Mat get_transistor_count() = 0;
     /* The number of cycles needed for the operation.
      * Similar to dynamic power as the value is stored after an operation */
     virtual int get_cycle_count() = 0;
