+++ include/simulator/base/pixel.h
@@ -36,9 +36,9 @@ class Pixel : public Component {
     void read(Register& reg);
     double last_frame_time();
 #ifdef TRACK_STATISTICS
-    cv::Mat& get_static_energy() override;
-    cv::Mat& get_dynamic_energy() override;
-    cv::Mat& get_transistor_count() override;
+    cv::Mat get_static_energy() override;
+    cv::Mat get_dynamic_energy() override;
+    cv::Mat get_transistor_count() override;
     void update_static(double time) override;
     int get_cycle_count() override;
     void print_stats(const CycleCounter& counter) override;
