+++ include/simulator/memory/dram3t_cell.h
@@ -36,9 +36,9 @@ public:
 #ifdef TRACK_STATISTICS
     void update_static(double time) override;
     int get_cycle_count() override;
-    cv::Mat& get_static_energy() override;
-    cv::Mat& get_dynamic_energy() override;
-    cv::Mat& get_transistor_count() override;
+    cv::Mat get_static_energy() override;
+    cv::Mat get_dynamic_energy() override;
+    cv::Mat get_transistor_count() override;
     void read(const cv::_InputOutputArray& mask) override;
     void read() override;
     void write(const cv::_InputOutputArray& mask) override;
