+++ include/simulator/memory/dram_array.h
@@ -50,9 +50,9 @@ class Dram : public Component {
     double fun_dynamic(const Config& config);
     void update_static(double time) override;
     int get_cycle_count() override;
-    cv::Mat& get_static_energy() override;
-    cv::Mat& get_dynamic_energy() override;
-    cv::Mat& get_transistor_count() override;
+    cv::Mat get_static_energy() override;
+    cv::Mat get_dynamic_energy() override;
+    cv::Mat get_transistor_count() override;
     void print_stats(const CycleCounter& counter) override;
 #endif
 };
