+++ src/simulator/adders/cla.cpp
@@ -64,15 +64,15 @@ void CarryLookAheadAdder::update_static(double time) {
     cv::add(this->array_static_energy_, this->static_power_ * time, this->array_static_energy_, this->internal_mask);
 }
 
-cv::Mat& CarryLookAheadAdder::get_static_energy() {
+cv::Mat CarryLookAheadAdder::get_static_energy() {
     return array_static_energy_;
 }
 
-cv::Mat& CarryLookAheadAdder::get_dynamic_energy() {
+cv::Mat CarryLookAheadAdder::get_dynamic_energy() {
     return array_dynamic_energy_;
 }
 
-cv::Mat& CarryLookAheadAdder::get_transistor_count() {
+cv::Mat CarryLookAheadAdder::get_transistor_count() {
     return array_transistor_count_;
 }
