+++ src/simulator/base/array.cpp
@@ -6,24 +6,25 @@
 
 #include <utility>
 
+Array::Array(int rows, int columns, Config& config, ProcessingElement pe) :
+    rows_(rows), columns_(columns),
+    config_(config), pe(std::move(pe))
+//    cla(rows, columns, 4, 4, 8, config),
+//    dram(rows, columns, 8, 8, 256, 1, 16, config)
+{}
+
 void Array::update_cycles(int cycles) {
 #ifdef TRACK_STATISTICS
     counter_ += cycles;
 #endif
 }
 
-Array::Array(int rows, int columns, Config& config, ProcessingElement pe) :
-    rows_(rows), columns_(columns),
-    config_(config), pe(std::move(pe)),
-    cla(rows, columns, 4, 4, 8, config),
-    dram(rows, columns, 8, 8, 256, 1, 16, config) {}
-
 #ifdef TRACK_STATISTICS
 
 void Array::update_static() {
     double time = (1.0 / config_.clock_rate) * counter_.get_cycles();
-    this->pe.update_static(counter_.get_cycles());
-    this->cla.update_static(time);
+    this->pe.update_static(time);
+//    this->cla.update_static(time);
 }
 
 void Array::print_stats() {
@@ -44,16 +45,16 @@ void Array::print_stats() {
     std::cout << "Cycle count: " << counter_.get_cycles() << "\n";
     double exec_time = ((double)counter_.get_cycles() / config_.clock_rate);
     std::cout << "Device execution time: " << exec_time << " s\n";
-    std::cout << "FPS: " << 500/exec_time << " \n";
+//    std::cout << "FPS: " << 500/exec_time << " \n";
     int pe_tc = cv::sum(this->pe.get_transistor_count())[0];
-    int cla_tc =  cv::sum(this->cla.get_transistor_count())[0];
-    int dram_tc = cv::sum(this->dram.get_transistor_count())[0];
-    int transistor_count =  pe_tc + cla_tc + dram_tc;
+//    int cla_tc =  cv::sum(this->cla.get_transistor_count())[0];
+//    int dram_tc = cv::sum(this->dram.get_transistor_count())[0];
+    int transistor_count =  pe_tc + 0 + 0;
     std::cout << "Array transistor count: " << transistor_count << "\n";
     std::cout << "Average transistor count per PE: " << transistor_count / (rows_ * columns_) << " \n";
-    double static_power = (cv::sum(this->pe.get_static_energy())[0] + cv::sum(this->cla.get_static_energy())[0]) / exec_time;
+    double static_power = (cv::sum(this->pe.get_static_energy())[0]) / exec_time;
     std::cout << "Array static power: " << static_power << " W\n";
-    double dynamic_power = (cv::sum(this->pe.get_dynamic_energy())[0] + cv::sum(this->cla.get_dynamic_energy())[0]) / exec_time;
+    double dynamic_power = (cv::sum(this->pe.get_dynamic_energy())[0]) / exec_time;
     std::cout << "Array dynamic power: " << dynamic_power << " W\n";
     std::cout << "Array total power: " << static_power + dynamic_power << " W\n";
