+++ src/simulator/memory/sram/sram6t_cell.cpp
@@ -38,15 +38,15 @@ double Sram6tCell::fun_dynamic_write(const Config& config) {
     return 6.305e-7;
 }
 
-cv::Mat& Sram6tCell::get_static_energy() {
+cv::Mat Sram6tCell::get_static_energy() {
     return array_static_energy_;
 }
 
-cv::Mat& Sram6tCell::get_dynamic_energy() {
+cv::Mat Sram6tCell::get_dynamic_energy() {
     return array_dynamic_energy_;
 }
 
-cv::Mat& Sram6tCell::get_transistor_count() {
+cv::Mat Sram6tCell::get_transistor_count() {
     return array_transistor_count_;
 }
